requirements.txt
black==22.12.0
flake8==6.0.0
kafka-python==2.0.2
mypy-extensions==0.4.3
requests==2.28.2
